making the code more modular
