+++ .gitignore
@@ -1,0 +1,3 @@
+bin/Graph
+.DS_Store
+Graph/.DS_Store
